añadir un if para que los botones de edicion y borrado del torneo solo salgan al iniciar sesion con el usuario que los creo

+++ backend/resources/views/torneo/index.blade.php
@@ -13,7 +13,7 @@
 
         <!-- Botón para crear un nuevo toarneo -->
         <div class="torneo-container">
-        <a href="torneos/creatorneo" class="btn btn-primary">Crear Torneo</a>
+            <a href="torneos/creatorneo" class="btn btn-primary">Crear Torneo</a>
         </div>
         <!-- Lista de torneos -->
         <div class="torneo-container">
@@ -32,9 +32,12 @@
                     @if ($usuarioRegistrado ?? '')
                         <a href="torneos/inscribir/{{ $torneo->id }}" class="btn btn-success">Inscribirse</a>
                     @endif
-                    <a href="torneos/editatorneo/{{ $torneo->id }}" class="btn btn-primary">Editar torneo</a>
-                    <a href="torneos/deletetorneo/{{ $torneo->id }}" class="btn btn-danger">Eliminar torneo</a>
+
                     <a href="torneos/muestratorneo/{{ $torneo->id }}" class="btn btn-info">Ver torneo</a>
+                    @if ($torneo->id_usu === Auth::id())
+                        <a href="torneos/editatorneo/{{ $torneo->id }}" class="btn btn-primary">Editar torneo</a>
+                        <a href="torneos/deletetorneo/{{ $torneo->id }}" class="btn btn-danger">Eliminar torneo</a>
+                    @endif
                     <!-- Botones de inscripción y ver participantes -->
                 </div>
             @endforeach
